requirements.txt
torch
torchvision
transformers
Pillow
faiss-cpu
streamlit
pandas
numpy
openai